+++ console/src/components/Services/Data/Common/Components/PrimaryKeySelector.js
@@ -46,11 +46,11 @@ const PrimaryKeySelector = ({ primaryKeys, columns, setPk, dispatch }) => {
 
       // show remove icon for all columns except last
       if (i + 1 === numPks) {
-        removeIcon = null;
+        removeIcon = <div className="ml-sm w-4" />;
       } else {
         removeIcon = (
           <i
-            className={`${styles.fontAwosomeClose} fa-lg fa fa-times`}
+            className="ml-sm w-4 fa fa-times cursor-pointer"
             data-test={`remove-pk-column-${i}`}
             onClick={dispatchRemove}
           />
@@ -58,10 +58,10 @@ const PrimaryKeySelector = ({ primaryKeys, columns, setPk, dispatch }) => {
       }
 
       return (
-        <div key={i} className={`form-group ${styles.pkEditorWrapper}`}>
+        <div key={i} className="flex items-center">
           <select
             value={pk || ''}
-            className={`${styles.select} ${styles.sample} form-control ${styles.add_pad_left}`}
+            className={`w-full form-control ${styles.add_pad_left}`}
             onChange={dispatchSet}
             data-test={`primary-key-select-${i}`}
           >
@@ -83,7 +83,7 @@ const PrimaryKeySelector = ({ primaryKeys, columns, setPk, dispatch }) => {
       );
     });
   };
-  return <div>{pkEditors()}</div>;
+  return <div className="space-y-md">{pkEditors()}</div>;
 };
 
 export default PrimaryKeySelector;