+++ console/src/components/Services/Data/Common/Components/UniqueKeySelector.js
@@ -1,5 +1,4 @@
 import React from 'react';
-import styles from '../../../../Common/TableCommon/Table.scss';
 
 const UniqueKeySelector = ({
   uniqueKeys,
@@ -55,16 +54,16 @@ const UniqueKeySelector = ({
 
     const removeIcon = (
       <i
-        className={`${styles.fontAwosomeClose} fa-lg fa fa-times`}
+        className="w-4 ml-sm fa fa-times cursor-pointer"
         data-test={`remove-uk-${index}-column-${i}`}
         onClick={removeUniqueCol}
       />
     );
 
     return (
-      <div key={i} className={`form-group ${styles.pkEditorWrapper}`}>
+      <div key={i} className="flex items-center">
         <select
-          className={`${styles.select} ${styles.sample} form-control ${styles.add_pad_left}`}
+          className="form-control"
           data-test={`unique-key-${index}-column-${i}`}
           value={uk}
           onChange={setUniqueCol}
@@ -83,9 +82,9 @@ const UniqueKeySelector = ({
   const newSelect = () => {
     const selectUniqueCol = e => dispatchSelectUniqueCol(e, numCols);
     return (
-      <div key={numCols} className={`form-group ${styles.pkEditorWrapper}`}>
+      <div className="flex" key={numCols}>
         <select
-          className={`${styles.select} ${styles.sample} form-control ${styles.add_pad_left}`}
+          className="form-control"
           data-test={`unique-key-${index}-column-${numCols}`}
           onChange={selectUniqueCol}
           value={''}
@@ -95,14 +94,20 @@ const UniqueKeySelector = ({
           </option>
           {getColumnOptions()}
         </select>
+        <div className="ml-sm w-4" />
       </div>
     );
   };
 
   return (
     <div>
-      {existingSelects}
-      {newSelect()}
+      <span className="flex items-center text-gray-600 font-semibold mb-formlabel">
+        Unique Keys:
+      </span>
+      <div className="space-y-md">
+        {existingSelects}
+        {newSelect()}
+      </div>
     </div>
   );
 };