+++ console/src/components/Services/Data/Function/Modify/SessionVarSection.tsx
@@ -46,19 +46,19 @@ const SessionVarSection: React.FC<SessionVarSectionProps> = ({
 
   return (
     <>
-      <hr className="my-md" />
-      <h4 className={styles.subheading_text}>
-        Session argument
+      <h4 className="flex items-center text-gray-600 font-semibold mb-formlabel">
+        Session Argument
         <ToolTip message="the function argument into which hasura session variables will be passed" />
         <KnowMoreLink href="https://hasura.io/docs/latest/graphql/core/schema/custom-functions.html#accessing-hasura-session-variables-in-custom-functions" />
       </h4>
       <div
-        className={isEditing ? styles.editorExpanded : styles.editorCollapsed}
+        className={
+          isEditing
+            ? 'space-y-md block rounded bg-white border border-gray-300 p-md mb-sm'
+            : 'space-y-md block'
+        }
       >
-        <div
-          className={`${styles.display_flex} ${styles.add_mar_bottom_mid}`}
-          data-test={`${functionName}-session-argument`}
-        >
+        <div className="mb-md" data-test={`${functionName}-session-argument`}>
           <Button
             className={styles.add_mar_small}
             color="white"
@@ -68,18 +68,17 @@ const SessionVarSection: React.FC<SessionVarSectionProps> = ({
           >
             {isEditing ? 'Close' : 'Edit'}
           </Button>
-          {configuration?.session_argument || 'No Session argument'}
         </div>
         {isEditing && (
           <>
             <EditorInput
-              label="Session argument"
+              label="Session Argument"
               value={sessVar}
               onChange={onSessVarChange}
               testID={functionName}
               placeholder="hasura_session"
             />
-            <div className={styles.add_mar_top_small}>
+            <div className="mt-md">
               <Button
                 type="submit"
                 color="yellow"