+++ console/src/components/Services/Data/TableBrowseRows/ViewTable.js
@@ -114,8 +114,6 @@ class ViewTable extends Component {
       throw new NotFoundError();
     }
 
-    const styles = require('../../../Common/Common.scss');
-
     // Is this a view
     const isView = !dataSource.isTable(tableSchema);
 
@@ -167,10 +165,12 @@ class ViewTable extends Component {
     let comment = null;
     if (tableSchema.comment) {
       comment = (
-        <div className={styles.add_mar_top}>
-          <div className={styles.commentText + ' alert alert-warning'}>
-            {tableSchema.comment}
-          </div>
+        <div
+          className={
+            'sm:max-w-xl rounded bg-secondary-light border border-gray-300 border-l-4 border-l-secondary py-sm px-md mt-lg'
+          }
+        >
+          {tableSchema.comment}
         </div>
       );
     }