+++ console/tailwind.config.js
@@ -31,33 +31,27 @@ module.exports = {
         },
         muted: {
           DEFAULT: '#475569',
-        }
+        },
       },
       fontFamily: {
         sans: ['Gudea', 'ui-sans-serif', 'system-ui'],
       },
       spacing: {
-        xs: '0.25rem',
-        sm: '0.5rem',
-        md: '1rem',
-        lg: '2rem',
-        xl: '2.75rem',
+        xs: '0.571rem',
+        sm: '0.857rem',
+        md: '1.429rem',
+        lg: '2.286rem',
+        xl: '2.857rem',
         input: '2.5rem',
         btn: '2.5rem',
-        btnsm: '1.75rem',
+        btnsm: '2rem',
+        formlabel: '0.571rem',
       },
       fontSize: {
-        sm: ['0.875rem', { lineHeight: '1.225rem' }],
-        base: ['1rem', { lineHeight: '1.4rem' }],
-        lg: ['1.25rem', { lineHeight: '1.75rem' }],
-        xl: ['1.75rem', { lineHeight: '2.45rem' }],
-      },
-      typography: {
-        DEFAULT: {
-          css: {
-            maxWidth: '100%',
-          },
-        },
+        sm: ['0.875rem', { lineHeight: '1.25rem' }],
+        base: ['1rem', { lineHeight: '1.5rem' }],
+        lg: ['1.143rem', { lineHeight: '1.429rem' }],
+        xl: ['1.714rem', { lineHeight: '2.143rem' }],
       },
     },
   },